% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVCBasicVaR.R
\docType{package}
\name{CVCBasicVaR}
\alias{CVCBasicVaR}
\alias{CVCBasicVaR-package}
\title{Caprari Varvus Cotugno Basic Value at Risk}
\description{
Download Data Efficiently and Compute Value at Risk with Parametric Approach
}
\details{
Package:                                              CVCBasicVaR \cr
Type:                   Package \cr
Version:                                    0.1.1 \cr
Date:                   2019-14-05 \cr
Depends:                                              R (>= 3.2.0), quantmod (>=0.4-13) \cr
Imports:                                    quantmod \cr
Lazyload:                                             yes \cr
License:                                    GPL-2 \cr
URL:                                        https://github.com/usiproject2019/CVCBasicVaR \cr
Repository:                                                \href{https://github.com/}{GitHub}

The goal of this package is to provide the tools to download prices and returns of stocks from
Yahoo finance, store them in a manageble way and, by using them, compute basic operations that concerns
Value at Risk.

\emph{What CVCBasicVaR is}

A useful tool to easy download prices and returns and organize them in a simple and efficient way,
with an intuitive visualization of the downloaded data. It provides also useful functions
to compute with the Parametric Approach the Value at Risk for single stocks or for an entire portfolio.

\emph{What CVCBasicVaR is not}

A replacement of \code{\link[quantmod:quantmod]{quantmod}}, the package on which the download functions
are based on, and a new Value at Risk approach. All the functions that provides the computation of
the Value at Risk are based on the well known standard formula. Nothing new was added.
}
\author{
Massimo Caprari, Anastasiya Varvus, Michele Cotugno (CVC)
}
