% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Price_Download.R
\name{Price_Download}
\alias{Price_Download}
\title{Download Daily Price or Volume from Yahoo Finance}
\usage{
Price_Download(ticker, price_type = 4, from, to)
}
\arguments{
\item{ticker}{a character vector of the yahoo finance ticker of the companies/indices for which prices/volume should be downloaded}

\item{price_type}{a number between 1 to 6 that specifies what is the base of the data that should be downloaded (default = 4).
1 = Daily price,
2 = Highest price,
3 = Lowest price,
4 = Closing price,
5 = Volume
6 = Adjusted price}

\item{from}{a class "date" element. It specifies the starting date from which the data will be downloaded (date of the first observation)}

\item{to}{a class "date" element. It specifies until when the data will be downloaded (date of the last observation)}
}
\value{
The function returns a class "xts" object that contains the downloaded prices/volume of each specified company in the specified time-interval
}
\description{
Download daily price or volume from yahoo finance for a single company or for a vector that contains
the tickers of several companies. Note that this function is based on the \code{\link[quantmod:getSymbols]{getSymbols}} function of the \code{\link[quantmod]{quantmod}}
package. Note that in order to have an omogeneous output all rows containing at least one missing observation will be deleted.
}
\examples{
## Not run:
Price_Download("^DJI", , "2019-05-01", "2019-05-10")


Stocks <- c("GE", "FORD", "BA", "^DJI", "DB")
Prices <- Price_Download(Stocks, 4, from = "2019-05-01", to = "2019-05-10")
## End (Not run)
}
\references{
\href{https://finance.yahoo.com/}{Yahoo Finance}
}
\seealso{
\code{\link[quantmod:getSymbols.yahoo]{getSymbols.yahoo}}
}
\author{
Massimo Caprari, Anastasiya Varvus, Michele Cotugno (CVC)
}
