% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Return_Download.R
\name{Return_Download}
\alias{Return_Download}
\title{Compute Daily Returns of Selected Stocks/Indices from Yahoo Finance}
\usage{
Return_Download(ticker, price_type = 4, from, to)
}
\arguments{
\item{ticker}{a character vector of the Yahoo Finance ticker of the companies/indices for which the daily return will be computed.}

\item{price_type}{a number that specifies what is the base of the data on which the returns will be computed (default = 4).
1 = Daily price,
2 = Highest price,
3 = Lowest price,
4 = Closing price,
6 = Adjusted price}

\item{from}{a class "date" element. It specifies the starting date on which the returns will be computed (date of the first observation). Note that, as the returns are calculated using the arithmetic approach, the first observation will be lost.}

\item{to}{a class "date" element. It specifies until when the returns will be computed (date of the last observation)}
}
\value{
a class "xts" element that contains the daily returns of each specified company in the specified time-interval
}
\description{
The function computes daily returns from Yahoo Finance for a single company or for a vector that contains
the tickers of several companies. Note that this function is based on the functions \code{\link[quantmod:getSymbols]{getSymbols}} and \code{\link[quantmod:dailyReturn]{dailyReturn}}  of the \code{\link[quantmod]{quantmod}}
package
}
\examples{
Return_Download("^DJI", price_type = 4, from="2019-05-01", to="2019-05-10")


Stocks <- c("GE", "FORD", "BA", "^DJI", "DB")
Returns <- Return_Download(Stocks, 4, from = "2019-05-01", to = "2019-05-10")

}
\references{
\href{https://finance.yahoo.com/}{Yahoo Finance}
}
\seealso{
\code{\link[quantmod:getSymbols.yahoo]{getSymbols.yahoo}}, \code{\link[TEST:Price_Download]{Price_Download}}, \code{\link[quantmod:dailyReturn]{dailyReturn}}
}
\author{
Massimo Caprari, Anastasiya Varvus, Michele Cotugno (CVC)
}
