% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rolling_Window_Volatility.R
\name{Rolling_Window_Volatility}
\alias{Rolling_Window_Volatility}
\title{Compute the Daily Volatility using a Rolling Window Approach}
\usage{
Rolling_Window_Volatility(return, window)
}
\arguments{
\item{return}{a class "xts" element: time series of returns. In order to get it, it is recommended to use \code{\link[CVCBasicVaR:Return_Download]{Return_Download}} function}

\item{window}{a class "numeric" element that specifies what is the time interval in days that will be used for the volatility computation}
}
\value{
The function returns a class "xts" object containing the volatility of the pre-specified assets for the pre-specified time window
It also returns some inputs of the function itself (window).
}
\description{
Using a rolling window approach, the function computes the daily volatility of
 a pre-specified vector of assets for given time series of returns.
}
\examples{
## Not Run:
Ticker <- c("GE","BA")
Returns <- Return_Download(myticker, 4, "2008-01-01", "2018-01-01")
Window <- c(10000)
Window_Volatility <- Rolling_Window_Volatility(myreturns, mywindow)

## End (Not run)

}
\references{
\href{https://finance.yahoo.com/}{Yahoo Finance}, \href{https://www.investopedia.com/terms/v/volatility.asp}{Investopedia}
}
\seealso{
\code{\link[quantmod:quantmod]{quantmod}}, \code{\link[quantmod:getSymbols.yahoo]{getSymbols.yahoo}}, \code{\link[CVCBasicVaR:Return_Download]{Return_Download}},
 \code{\link[quantmod:dailyReturn]{dailyReturn}}
}
\author{
Massimo Caprari, Anastasiya Varvus, Michele Cotugno (CVC)
}
