% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_VaR.R
\name{Single_VaR}
\alias{Single_VaR}
\title{Compute the Value at Risk for Single Positions}
\usage{
Single_VaR(return, confidence_level, wealth, T)
}
\arguments{
\item{return}{a class "xts" element: time serie of returns. In order to get it, it is recommended to use \code{\link[TEST:Return_Download]{Return_Download}} function}

\item{confidence_level}{a number between 0 and 1 that specifies what is the confidence level to use in the computation}

\item{wealth}{a numeric vector that specifies the monetary exposure assumed in every stock/index (note that the order of elements of the vector must reflect the order of the elements of the returns)}

\item{T}{a number that specifies the time interval (in days) based on which the Value at Risk is computed}
}
\description{
Using the Asset Normal Approach, the function computes the Value at Risk for a given time series of return and vector of positions.
}
\examples{
Stocks <- c("GE", "AAPL", "BA", "DB", "FORD", "^DJI")
Returns <- Return_Download(Stocks, 4, from = Sys.Date()-365.25*3, to = Sys.Date())
Wealth <- c(100000, 50000, 130000, 75000, 500000, 930000)
IndividualVaR <- SingleVaR(Returns, 0.99, Wealth, 10)

}
\references{
\href{https://finance.yahoo.com/}{Yahoo Finance}
}
\seealso{
\code{\link[quantmod:quantmod]{quantmod}}, \code{\link[quantmod:getSymbols.yahoo]{getSymbols.yahoo}}, \code{\link[TEST:Return_Download]{Return_Download}},
 \code{\link[quantmod:dailyReturn]{dailyReturn}},
}
\author{
Massimo Caprari, Anastasiya Varvus, Michele Cotugno (CVC)
}
