% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_VaR_Rolling_Window.R
\name{Single_VaR_Rolling_Window}
\alias{Single_VaR_Rolling_Window}
\title{Compute the Value at Risk for Single Positions using a Rolling Window Approach}
\usage{
Single_VaR_Rolling_Window(price, wealth, reference_date, window,
  confidence_level, T)
}
\arguments{
\item{price}{a class "xts" element: time serie of prices In order to get it, it is recommended to use \code{\link[CVCBasicVaR:Price_Download]{Price_Download}} function}

\item{wealth}{a numeric vector that specifies the monetary exposure assumed in every stock/index (note that the order of elements of the vector must reflect the order of the elements of the prices)}

\item{reference_date}{a class "date" element, it is the date on which the wealth is referred. It is used to compute the number of stocks for each assets.}

\item{window}{a class "numeric" element that specifies what is the time interval in days that will be used for the volatility computation}

\item{confidence_level}{a number between 0 and 1 that specifies what is the confidence level to use in the computation}

\item{T}{a number that specifies the time interval (in days) based on which the Value at Risk is computed}
}
\value{
The function returns a class "xts" object that contains the daily VaRs of every assets for the its relative monetary exposure in a certain date
and for a chosen window of observation.
It also returns some inputs of the function itself (Confidence level, time interval).
}
\description{
Using the Asset Normal Approach, the function computes the Value at Risk each day for different assets
using a pre-specified time window for the volatility computation and a certain vector of monetary exposure in each stock/index
in a given day.
}
\examples{
## Not Run:
Stocks <- c("GE", "AAPL")
Price <- Price_Download(Stocks, 6, from = "2010-01-04", to = "2018-01-01")
Wealth <- c(100000, 50000)
RollingVaR <- Single_VaR_Rolling_Window(Price, Wealth, "2010-01-02", 250, 0.99, 1)
## End (Not run)
}
\references{
\href{https://finance.yahoo.com/}{Yahoo Finance}, \href{ https://www.investopedia.com/terms/r/rollingreturns.asp}{Investopedia}
}
\seealso{
\code{\link[quantmod:quantmod]{quantmod}}, \code{\link[quantmod:getSymbols.yahoo]{getSymbols.yahoo}}, \code{\link[CVCBasicVaR:Price_Download]{Price_Download}},
 \code{\link[CVCBasicVaR:Rolling_Window_Volatility]{Rolling_Window_Volatility}}
}
\author{
Massimo Caprari, Anastasiya Varvus, Michele Cotugno (CVC)
}
